Added bugs
I dont remember what i was doing at the start of this update,
But most of the time was spent fixing a bug that appeared just because. I havent touched any parts of the game that could potentioally lead to this bug.
Also I havent even found a solution, all I could do is figure out how to set up a crutch.
Also added a new brick type.